exersize 1 added tests
