+++ tdd-oop-inheritance.CSharp.Test/ArticleTest.cs
@@ -1,41 +1,40 @@
 ﻿using tdd_oop_inheritance.CSharp.Main;
 using NUnit.Framework;
 
-namespace tdd_oop_inheritance.CSharp.Test
+namespace tdd_oop_inheritance.CSharp.Test;
+
+class ArticleTest
 {
-    class ArticleTest
-    {
-        [Test]
-        public void shouldCheckOutIfAvailable()
-            {
-                Article article = new Article("JUnit Rocks");
-                Assert.AreEqual("item has been checked out", article.checkOut());
-            }
+    [Test]
+    public void shouldCheckOutIfAvailable()
+        {
+            Article article = new Article("JUnit Rocks");
+            Assert.AreEqual("item has been checked out", article.checkOut());
+        }
 
-        [Test]
-        public void shouldDeclineIfNotAvailableToCheckout()
-            {
-                Article article = new Article("JUnit Rocks");
-                article.checkOut();
+    [Test]
+    public void shouldDeclineIfNotAvailableToCheckout()
+        {
+            Article article = new Article("JUnit Rocks");
+            article.checkOut();
 
-                Assert.AreEqual("item is currently on loan", article.checkOut());
-            }
+            Assert.AreEqual("item is currently on loan", article.checkOut());
+        }
 
-        [Test]
-        public void shouldCheckInIfOnLoan()
-            {
-                Article article = new Article("JUnit Rocks");
-                article.checkOut();
+    [Test]
+    public void shouldCheckInIfOnLoan()
+        {
+            Article article = new Article("JUnit Rocks");
+            article.checkOut();
 
-                Assert.AreEqual("item has been checked in", article.checkIn());
-            }
+            Assert.AreEqual("item has been checked in", article.checkIn());
+        }
 
-        [Test]
-        public void shouldDeclineCheckInIfNotOnLoan()
-            {
-                Article article = new Article("JUnit Rocks");
+    [Test]
+    public void shouldDeclineCheckInIfNotOnLoan()
+        {
+            Article article = new Article("JUnit Rocks");
 
-                Assert.AreEqual("item is not currently on loan", article.checkIn());
-            }
-    }
+            Assert.AreEqual("item is not currently on loan", article.checkIn());
+        }
 }
